requirements.txt
Flask==1.1.2
python-telegram-bot==12.8
beautifulsoup4
configparser
requests